Arreglado el reload automatico y detalle sobre mensaje de Bienvenid@,[nombre]
